Move RecordingNotifier to recordings folder
