+++ Jellyfin.Server/Startup.cs
@@ -6,9 +6,9 @@ using System.Net.Mime;
 using System.Text;
 using Emby.Server.Implementations.EntryPoints;
 using Jellyfin.Api.Middleware;
-using Jellyfin.LiveTv;
 using Jellyfin.LiveTv.EmbyTV;
 using Jellyfin.LiveTv.Extensions;
+using Jellyfin.LiveTv.Recordings;
 using Jellyfin.MediaEncoding.Hls.Extensions;
 using Jellyfin.Networking;
 using Jellyfin.Networking.HappyEyeballs;
